[Pdm2-4] Extract Form from LoginContent
